requirements.txt
boto3==1.34.54
pymysql==1.1.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0
pytest==7.4.4
pytest-dotenv==0.5.2
pytest-cov==4.1.0
psycopg2-binary==2.9.9